Add last remaining operator tests
