+++ test/operator.jl
@@ -1,6 +1,25 @@
 using JuMP
 using Compat
 using Compat.Test
+using OffsetArrays
+
+# For "DimensionMismatch when performing vector-matrix multiplication with custom types #988"
+import Base: +, *
+struct MyType{T}
+    a::T
+end
+struct MySumType{T}
+    a::T
+end
+Base.one(::Type{MyType{T}}) where {T} = MyType(one(T))
+Base.zero(::Type{MySumType{T}}) where {T} = MySumType(zero(T))
+Base.zero(::MySumType{T}) where {T} = MySumType(zero(T))
+Base.transpose(t::MyType) = MyType(t.a)
+Base.transpose(t::MySumType) = MySumType(t.a)
++(t1::MyT, t2::MyS) where {MyT<:Union{MyType, MySumType}, MyS<:Union{MyType, MySumType}} = MySumType(t1.a+t2.a)
+*(t1::MyType{S}, t2::T) where {S, T} = MyType(t1.a*t2)
+*(t1::S, t2::MyType{T}) where {S, T} = MyType(t1*t2.a)
+*(t1::MyType{S}, t2::MyType{T}) where {S, T} = MyType(t1.a*t2.a)
 
 @testset "Operator" begin
     @testset "Basic operator overloads" begin
@@ -527,6 +546,65 @@ using Compat.Test
             @test vec_eq(f7, mat)
             @test map(c -> c.set.lower, c7) == -[1:3;]
             @test all(c -> c.set.upper == 3, c7)
+        end
+    end
+
+    @testset "Operators for non-Array AbstractArrays" begin
+        m = Model()
+        @variable(m, x[1:3])
+
+        # This is needed to compare arrays that have nonstandard indexing
+        elements_equal(A::AbstractArray{T, N}, B::AbstractArray{T, N}) where {T, N} = all(a == b for (a, b) in zip(A, B))
+
+        for x2 in (OffsetArray(x, -length(x)), view(x, :), sparse(x))
+            @test elements_equal(+x, +x2)
+            @test elements_equal(-x, -x2)
+            @test elements_equal(x .+ first(x), x2 .+ first(x2))
+            @test elements_equal(x .- first(x), x2 .- first(x2))
+            @test elements_equal(first(x) .- x, first(x2) .- x2)
+            @test elements_equal(first(x) .+ x, first(x2) .+ x2)
+            @test elements_equal(2 .* x, 2 .* x2)
+            @test elements_equal(first(x) .+ x2, first(x2) .+ x)
+            @test sum(x) == sum(x2)
+            if !JuMP.one_indexed(x2)
+                @test_throws DimensionMismatch x + x2
+            end
+        end
+    end
+
+    @testset "diagm for non-Array AbstractArrays" begin
+        m = Model()
+        @variable(m, x[1:3])
+
+        for x2 in (OffsetArray(x, -length(x)), view(x, :), sparse(x))
+            if !JuMP.one_indexed(x2)
+                @test_throws AssertionError diagm(x2)
+            else
+                @test diagm(x) == diagm(x2)
+            end
+        end
+    end
+
+    @testset "DimensionMismatch when performing vector-matrix multiplication with custom types #988" begin
+        m = Model()
+        @variable m Q[1:3, 1:3] PSD
+
+        x = [MyType(1), MyType(2), MyType(3)]
+        y = Q * x
+        z = x' * Q
+        ElemT = MySumType{AffExpr}
+        @test y isa Vector{ElemT}
+        @test size(y) == (3,)
+        @test z isa RowVector{ElemT, ConjArray{ElemT, 1, Vector{ElemT}}}
+        @test size(z) == (1, 3)
+        for i in 1:3
+            # Q is symmetric
+            a = zero(AffExpr)
+            a += Q[1,i]
+            a += 2Q[2,i]
+            a += 3Q[3,i]
+            # Q[1,i] + 2Q[2,i] + 3Q[3,i] is rearranged as 2 Q[2,3] + Q[1,3] + 3 Q[3,3]
+            @test z[i].a == y[i].a == a
         end
     end
 end
